requirements.txt
aenum==3.1.0
aiohttp==3.7.4.post0
aiohttp-jinja2==1.5
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.5
grpcio==1.40.0
grpcio-tools==1.40.0
idna==3.2
Jinja2==3.0.1
MarkupSafe==2.0.1
multidict==5.1.0
protobuf==3.18.0
python-can==3.3.4
requests==2.26.0
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.6
wrapt==1.12.1
yarl==1.6.3
